EventSourcedProjection, API entry point based on Source
* EventSourcedProjection for eventsByTag sources
